refactor(profile): 处理 Sourcery code review 反馈
- fetchReposPage 增加 visibility=public：原 affiliation=owner 默认带回私有
  仓库，会让列表条数大于头部 publicRepos 计数；显式过滤为公开仓库以对齐口径
- 抽取 PagedListViewModel<T> 泛型分页基类，GithubUserListViewModel /
  RepoListViewModel 收敛为薄子类（仅提供 fetchPage / keyOf / pageSize）
- 抽取 PaginatedListScaffold 通用列表骨架（TopAppBar + loading/error/empty/
  列表四态 + 触底翻页），两个列表 Screen 复用，消除重复
- 列表页返回按钮补 contentDescription（用现成 Res.string.back），与
  Settings/Favorites/Feedback 等既有页面的无障碍写法一致
